test(github-client): ETagキャッシュテストを別ファイルへ分離
【背景】test_github_client.pyが肥大化していたため、ETagキャッシュ関連テストを
別ファイルへ分離する分割作業(U5)の一環として実施。

【変更内容】ETagキャッシュヒット/ミス、304レスポンス、キャッシュキー生成、
max_cache_entries検証など10関数をtest_github_client_etag.pyへ移動。

品質ゲート: 1456 tests passed / 97.60% coverage / ruff 0 errors / mypy 0 errors
